[JT-28] refactor: 패키지 구조 변경, member entity 필수 필드만 validation 동작하도록 변경
