+++ module-domain/src/main/java/com/devtoon/jtoon/global/common/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.devtoon.jtoon.common;
+package com.devtoon.jtoon.global.common;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
@@ -19,7 +19,7 @@ import lombok.NoArgsConstructor;
 public abstract class BaseTimeEntity {
 
 	@CreatedDate
-	@Column(name = "created_at", updatable = false)
+	@Column(name = "created_at", updatable = false, nullable = false)
 	private LocalDateTime createdAt;
 
 	@LastModifiedDate
