+++ module-domain/src/main/java/com/devtoon/jtoon/global/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.devtoon.jtoon.common;
+package com.devtoon.jtoon.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
