+++ module-domain/src/test/java/com/devtoon/jtoon/annotation/JpaHelper.java
@@ -4,7 +4,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabas
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
-import com.devtoon.jtoon.common.JpaConfig;
+import com.devtoon.jtoon.global.config.JpaConfig;
 
 @DataJpaTest
 @Import(JpaConfig.class)
